Identifying spatial interdependence in panel data with large N and small T
This paper develops a simple two-stage variational Bayesian algorithm to estimate panel spatial autoregressive models, where   the number of cross-sectional units, is much larger than   the number of time periods without restricting the spatial effects using a predetermined weighting matrix. We use Dirichlet-Laplace priors for variable selection and parameter shrinkage. Without imposing any a priori structures on the spatial linkages between variables, we let the data speak for themselves. Extensive Monte Carlo studies show that our method is super-fast and our estimated spatial weights matrices strongly resemble the true spatial weights matrices. As an illustration, we investigate the spatial interdependence of European Union regional gross value added growth rates. In addition to a clear pattern of predominant country clusters, we have uncovered a number of important between-country spatial linkages which are yet to be documented in the literature. This new procedure for estimating spatial effects is of particular relevance for researchers and policy makers alike.